core: split force initialization and verlet-list build into own translation units
Backport of 3a54ec10ee to the python branch: same TU layout and
update_verlet_state wrapper, but the moved function bodies are the python
branch's own (the source branch's bodies are ParticleStore column kernels).
Pure code motion relative to the python branch.
