+++ src/core/energy.cpp
@@ -34,6 +34,7 @@
 #include "nonbonded_interactions/nonbonded_interaction_data.hpp"
 #include "short_range_cabana.hpp"
 #include "short_range_loop.hpp"
+#include "short_range_verlet.hpp"
 #include "system/GpuParticleData.hpp"
 #include "system/System.hpp"
 
@@ -86,7 +87,7 @@ std::shared_ptr<Observable_stat> System::calculate_energy() {
                                            coulomb.cutoff(),
                                            dipoles.cutoff(),
                                            inactive_cutoff};
-  update_cabana_state(*cell_structure, verlet_criterion,
+  update_verlet_state(*cell_structure, verlet_criterion,
                       get_interaction_range(), propagation->integ_switch);
 
   EnergyBinLayout layout{
