+++ src/core/forces.cpp
@@ -32,6 +32,7 @@
 #include "communication.hpp"
 #include "constraints/Constraints.hpp"
 #include "electrostatics/icc.hpp"
+#include "forces_init.hpp"
 #include "forces_inline.hpp"
 #include "galilei/ComFixed.hpp"
 #include "immersed_boundary/ImmersedBoundaries.hpp"
@@ -44,10 +45,10 @@
 #include "rotation.hpp"
 #include "short_range_cabana.hpp"
 #include "short_range_loop.hpp"
+#include "short_range_verlet.hpp"
 #include "system/GpuParticleData.hpp"
 #include "system/System.hpp"
 #include "thermostat.hpp"
-#include "thermostats/langevin_inline.hpp"
 #include "virtual_sites/com.hpp"
 #include "virtual_sites/relative.hpp"
 
@@ -66,69 +67,6 @@
 #include <memory>
 #include <span>
 #include <variant>
-
-/** External particle forces */
-static ParticleForce external_force(Particle const &p) {
-  ParticleForce f = {};
-
-#ifdef ESPRESSO_EXTERNAL_FORCES
-  f.f += p.ext_force();
-#ifdef ESPRESSO_ROTATION
-  f.torque += p.ext_torque();
-#endif
-#endif
-
-#ifdef ESPRESSO_ENGINE
-  // apply a swimming force in the direction of
-  // the particle's orientation axis
-  if (p.swimming().swimming and !p.swimming().is_engine_force_on_fluid) {
-    f.f += p.swimming().f_swim * p.calc_director();
-  }
-#endif
-
-  return f;
-}
-
-/** Combined force initialization and Langevin noise application */
-static void init_forces_and_thermostat(System::System const &system) {
-#ifdef ESPRESSO_CALIPER
-  CALI_CXX_MARK_FUNCTION;
-#endif
-
-  auto &cell_structure = *system.cell_structure;
-  auto const &propagation = *system.propagation;
-  auto const &thermostat = *system.thermostat;
-  auto const kT = thermostat.kT;
-  auto const time_step = system.get_time_step();
-
-  // Check if Langevin thermostat is active
-  bool const langevin_active =
-      thermostat.langevin &&
-      (propagation.used_propagations &
-       (PropagationMode::TRANS_LANGEVIN | PropagationMode::ROT_LANGEVIN));
-
-  // Single pass over all local particles
-  cell_structure.for_each_local_particle([&](Particle &p) {
-    // Initialize force with external forces
-    p.force_and_torque() = external_force(p);
-
-    // Apply Langevin noise if thermostat is active
-    if (langevin_active) {
-      auto const &langevin = *thermostat.langevin;
-      if (propagation.should_propagate_with(p, PropagationMode::TRANS_LANGEVIN))
-        p.force() += friction_thermo_langevin(langevin, p, time_step, kT);
-#ifdef ESPRESSO_ROTATION
-      if (propagation.should_propagate_with(p, PropagationMode::ROT_LANGEVIN))
-        p.torque() += convert_vector_body_to_space(
-            p, friction_thermo_langevin_rotation(langevin, p, time_step, kT));
-#endif
-    }
-  });
-  cell_structure.reset_local_force_and_torque();
-
-  // Initialize ghost forces (unchanged)
-  cell_structure.ghosts_reset_forces();
-}
 
 static void force_capping(CellStructure &cell_structure, double force_cap) {
   if (force_cap > 0.) {
@@ -311,7 +249,7 @@ void System::System::calculate_forces() {
                                            dipoles.cutoff(),
                                            collision_detection_cutoff};
 
-  update_cabana_state(*cell_structure, verlet_criterion,
+  update_verlet_state(*cell_structure, verlet_criterion,
                       get_interaction_range(), propagation->integ_switch);
 #ifdef ESPRESSO_ELECTROSTATICS
   if (coulomb.impl->extension) {
